# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workflow Service",
...
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://windmill.distribute.org"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "nullable": true
          }
        },
        "required": [
          "error"
        ]
      },
...
      "WorkflowRunResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Run UUID. Use this to poll status via GET /workflow-runs/{id}."
          },
          "workflowId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "orgId": {
            "type": "string"
          },
          "campaignId": {
            "type": "string",
            "nullable": true
          },
          "subrequestId": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true,
            "description": "User ID from the execution context."
          },
          "runId": {
            "type": "string",
            "nullable": true,
            "description": "Runs-service run ID for this execution (auto-created)."
          },
          "windmillJobId": {
            "type": "string",
            "nullable": true,
            "description": "Internal Windmill job ID (managed automatically)."
          },
          "windmillWorkspace": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "Run status: queued, running, completed, failed, or cancelled."
          },
          "inputs": {
            "nullable": true,
            "description": "The inputs that were passed at execution time."
          },
          "result": {
            "nullable": true,
            "description": "Workflow result (available when status is completed). Contains the output of the last node."
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "Error message (available when status is failed)."
          },
          "startedAt": {
            "type": "string",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "workflowId",
          "orgId",
          "campaignId",
          "subrequestId",
          "userId",
          "runId",
          "windmillJobId",
          "windmillWorkspace",
          "status",
          "error",
          "startedAt",
          "completedAt",
          "createdAt"
        ]
      },
      "ExecuteWorkflowRequest": {
        "type": "object",
        "properties": {
          "inputs": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "Runtime inputs for the workflow. Accessible in nodes via $ref:flow_input.fieldName."
          }
        }
      },
      "WorkflowRunDebugResponse": {
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "windmillJobId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "flowStatus": {
            "nullable": true,
            "description": "Windmill flow_status object. Contains per-module execution details including resolved inputs, outputs, and timing for each step in the flow."
          },
          "result": {
            "nullable": true,
            "description": "Final flow result."
          }
        },
        "required": [
          "runId",
          "windmillJobId",
          "status"
        ]
      },
...
      "ExecuteByNameRequest": {
        "type": "object",
        "properties": {
          "orgId": {
            "type": "string",
            "minLength": 1,
            "description": "Deprecated — ignored for workflow lookup. The executing org is identified via the x-org-id header. Kept optional for backwards compatibility."
          },
          "inputs": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "Runtime inputs for the workflow. Accessible in nodes via $ref:flow_input.fieldName."
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/workflows/{id}/execute": {
      "post": {
        "summary": "Execute a workflow",
        "tags": [
          "Workflow Runs"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteWorkflowRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Execution started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workflow-runs/{id}": {
      "get": {
        "summary": "Get a workflow run",
        "description": "Returns the current status and result of a workflow execution. If still running, polls the engine for the latest status before responding.",
        "tags": [
          "Workflow Runs"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workflow-runs/{id}/debug": {
      "get": {
        "summary": "Debug a workflow run",
        "description": "Returns per-step execution details from the Windmill engine, including resolved inputs and outputs for each module. Use this to diagnose runtime issues that aren't visible in the final result.",
        "tags": [
          "Workflow Runs"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Debug details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunDebugResponse"
                }
              }
            }
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workflow-runs": {
      "get": {
        "summary": "List workflow runs",
        "tags": [
          "Workflow Runs"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "workflowId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "orgId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "campaignId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of runs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "workflowRuns": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WorkflowRunResponse"
                      }
                    }
                  },
                  "required": [
                    "workflowRuns"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/workflow-runs/{id}/cancel": {
      "post": {
        "summary": "Cancel a workflow run",
        "tags": [
          "Workflow Runs"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Run cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/workflows/by-name/{name}/execute": {
      "post": {
        "summary": "Execute a workflow by name",
        "description": "Starts an async execution of a previously deployed workflow. Returns a run object with an ID — poll GET /workflow-runs/{id} for status (queued → running → completed/failed). Pass runtime data via inputs (accessible in nodes as $ref:flow_input.fieldName).",
        "tags": [
          "Workflow Runs"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "name",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteByNameRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Execution started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workflow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}